% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_allyrs.R
\name{alloc_allyrs}
\alias{alloc_allyrs}
\title{An optimised demand allocation model}
\usage{
alloc_allyrs(A, cohort, scenario1, scenario2, scenario3, scenario4, scenario5,
  scenario6, scenario7, scenario8, scenario9, scenario10)
}
\arguments{
\item{clustername, }{cohort, scenarios(1 to N)}
}
\description{
allocate students to schools based on capacity and travel time
simulate for all years
}
\examples{
alloc_yrs()
}
\keyword{asset}
\keyword{planner}

