\name{clstrplnr-package}
\alias{clstrplnr-package}
\alias{clstrplnr}
\docType{package}
\title{
\packageTitle{clstrplnr}
}
\description{
\packageDescription{clstrplnr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{clstrplnr}
\packageIndices{clstrplnr}
~~ indices - update ~~
}
\author{
\packageAuthor{clstrplnr}

Maintainer: \packageMaintainer{clstrplnr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
